requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.21.3
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
dataclasses-json==0.6.6
datasets==2.14.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
docstring-parser==0.15
dspy-ai==2.4.9
edgartools==2.21.1
fastapi==0.110.3
fastcore==1.5.38
filelock==3.14.0
financial-datasets==0.1.14
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2023.10.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
humanize==4.9.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
instructor==0.6.8
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.2.0
langchain-chroma==0.1.1
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.60
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.30.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optuna==3.6.1
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.5.0
py==1.11.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.1
regex==2024.5.15
requests==2.32.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
structlog==24.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
textdistance==4.6.2
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
